ml_services/prediction_service/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
# Указываем ссылку на CUDA 12.1 версию PyTorch, чтобы GPU точно работал
torch --index-url https://download.pytorch.org/whl/cu121